refactor: update backup handling to use structured error responses and improve type imports
