+++ src/backend/features/device/renameDevice.ts
@@ -1,5 +1,5 @@
 import { logger } from '@internxt/drive-desktop-core/build/backend';
-import { Device } from '../../../apps/main/device/service';
+import { Device } from '../backup/types/Device';
 import { Result } from '../../../context/shared/domain/Result';
 import { driveServerModule } from '../../../infra/drive-server/drive-server.module';
 import { getDeviceIdentifier } from './getDeviceIdentifier';
